Add density charts view with DRY schema refactor

+++ src/schemas/common/error.schema.ts
@@ -2,7 +2,7 @@ import { z } from 'zod'
 
 // Generic error schema - matches Fastify Sensible HttpError structure
 export const ErrorSchema = z.object({
-  statusCode: z.number(),
+  statusCode: z.number().int(),
   code: z.string(),
   error: z.string(),
   message: z.string().min(1),
